+++ src/Raven.CodeAnalysis/Workspaces/CodeFixes/BuiltInCodeRefactoringProviders.cs
@@ -1,0 +1,8 @@
+using System.Collections.Immutable;
+
+namespace Raven.CodeAnalysis;
+
+public static class BuiltInCodeRefactoringProviders
+{
+    public static ImmutableArray<CodeRefactoringProvider> CreateDefault() => [];
+}